Add IVersionedDatastore hook for iiif record augmentation
Using the new datastore_multisearch_modify_response hook we add a iiif
key to the root of the record and build the manifest straight into it.
This is currently only implemented for the vfactor iiif work,
generification to come.

Doing this is useful when a user is trying to get the manifests for
multiple records as calling the manifest endpoint for each record is
inefficient.