Stop adding /info.json onto the end of the image body ID
This is incorrect. I've seen manifests that point to the full/max/0/default.jpg version but I think given our base image endpoints at /media/{id} show an image this makes sense. If you get the manifest and you want to get the info.json (or indeed any IIIF iamge data) you just add it to the end of the URL which is nice.