fix: stop producing manifests if no images are available on a record
Previously, when the manifest was generated a record with no images would produce a manifest with no canvas items. The IIIF presentation spec says that you must have an items in your manifest but it doesn't specify whether there have to be objects within it. I think a manifest with no items is a little silly so my interpretation is to avoid generating empty manifests. Additionally, this means that callers can find out whether a record has a corresponding manifest that it is worth showing users based on whether the manifest builder returns a manifest or not. Now if it doesn't return a manifest you know there are no images available to build one.