fix: allow the images in a manifest to be correctly opened in mirador etc
Without this service section, Mirador (and others) has no idea that the image is a IIIF image and therefore just loads the static image. This change fixes that problem, but will break any non-IIIF image as a consequence. Fortunately, we're only using this functionality for IIIF specimen images so this large oversight isn't a problem right now. We will need to fix this later by finding a way to conditionally include or exclude this service section depending on whether the image is a IIIF image or not.